Bugfix for plotting.
Somehow providing `plot_analysis` with the constructed `DataFrame` results in an
error when calling `~df['gpu']`. This does _not_ happen when calling
`(~df['gpu']) & (df['host'])`. It seems to be related to the way we create the
`DataFrame`. Reading the previously generated CSV and providing this `DataFrame`
works without any problems.