Add 5 minutes of run time to every requested job.
When requesting the same run time as the actual run time specified for GROMACS
with the `-maxh` parameter, it's possible that the queuing system kills the job
prematurely. GROMACS will not have the time to finish the actual job and output
the Performance information that the `benchmark analyze` script needs.

Fixes #5.